Fixed error where top 16 bits of lba48 were not used in ata_io_dma(). Added cast to uint64_t for block number in detect_filesystem(), fat_getstartblock(), fat_detectchange(). Changed block,blockno and rb parameter in fd_io() call in fat_rename(), fat_create_int(), fat_read, fat_write(),fat_findfreeblock(), fat_getstartblock() from size_t to uint64_t. Changed currentblock and findblock fields in FILERECORD struct from size_t to uint64_t. Changed count and countx variables from size_t to uint64_t in fat_findfreeblock(). Changed block parameter in fat_readlfn(), fat_writelfn, fat_deletelfn(), fat_createlfn(), far_getnextblock() from size_t to uint64_t. Changed count loop variable to uint64_t block_write_count variable in deletelfn().

+++ filemanager/vfs.c
@@ -25,6 +25,7 @@
 #include "../processmanager/mutex.h"
 #include "../devicemanager/device.h"
 #include "../header/bootinfo.h"
+#include "../header/debug.h"
 
 size_t findfirst(char *name,FILERECORD *buf);
 size_t findnext(char *name,FILERECORD *buf);
@@ -1033,7 +1034,7 @@ lock_mutex(&vfs_mutex);
 blockbuf=kernelalloc(VFS_MAX);
 if(blockbuf == NULL) return(-1);
 
-if(blockio(0,drive,0,blockbuf) == -1) return(-1);		/* read block */
+if(blockio(0,drive,(uint64_t) 0,blockbuf) == -1) return(-1);		/* read block */
 
 /* check magic number */
